updating bottom navigation state to navigate to respective pagegs
