+++ lib/classes/messages.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class MessagesScreen extends StatelessWidget {
+  const MessagesScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Messages Screen'),
+    );
+  }
+}
