+++ lib/classes/settings.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class SettingsScreen extends StatelessWidget {
+  const SettingsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Settings Screen'),
+    );
+  }
+}
