+++ lib/screens/loginscreen.dart
@@ -92,7 +92,7 @@ class LoginPage extends StatelessWidget {
                       Navigator.pushReplacement(
                         context,
                         MaterialPageRoute(
-                          builder: (context) => const HomePage(),
+                          builder: (context) => HomePage(),
                         ),
                       );
                     } on FirebaseAuthException catch (e) {
